Added ability to write custom parsers for new api types Separated base apipie parser logic into a new class
